adding the review comments
